# Use VecDeque instead of Vec in sigverify stage