#### Problem

Vec has bad remove(0) performance.

#### Summary of Changes

Avoid bad performance of remove(0) for a single sender and use VecDeque instead.

Before:
```
test bench_packet_discard_many_senders  ... bench:   1,943,585 ns/iter (+/- 3,441)
test bench_packet_discard_single_sender ... bench:  55,018,780 ns/iter (+/- 326,300)
```

After:
```
test bench_packet_discard_many_senders  ... bench:   1,894,268 ns/iter (+/- 12,025)
test bench_packet_discard_single_sender ... bench:   1,361,563 ns/iter (+/- 8,583)
```

Fixes #